better organization for modularity while still being able to reuse key functions and fixtures between test folders and modules
