+++ tests/pipeline/test_multi_pipeline_sample.py
@@ -3,127 +3,13 @@
 import os
 import time
 import pytest
-from pypiper import Stage
 from pypiper.utils import checkpoint_filepath
-from tests.helpers import fetch_checkpoint_files, SafeTestPipeline
+from tests.helpers import fetch_checkpoint_files
+from .conftest import get_pipeline
 
 
 __author__ = "Vince Reuter"
 __email__ = "vreuter@virginia.edu"
-
-
-READ_ALIGNER_FILENAME = "aligner.lst"
-PEAK_CALLER_FILENAME = "caller.lst"
-
-
-
-def pytest_generate_tests(metafunc):
-    """ Dynamic test case parameterization for this module. """
-    if "pl_name" in metafunc.fixturenames:
-        metafunc.parametrize(
-            "pl_name", [read_aligner.__name__, call_peaks.__name__])
-
-
-
-# Dummy functions used as elements of pipeline stages() collections.
-def merge_input():
-    pass
-
-def qc():
-    pass
-
-def align_reads():
-    pass
-
-def call_peaks():
-    pass
-
-
-
-class FunctionNameWriterPipeline(SafeTestPipeline):
-    """ Basic pipeline that writes to file the names of its functions. """
-    
-    
-    def __init__(self, name, outfolder, filename, functions):
-        """
-        Name and outfolder go to generic pipeline ctor; filename and functions
-        collection are used specifically by instances of this class.
-
-        :param str name: Name for this pipeline.
-        :param str outfolder: Path to pipeline's output folder.
-        :param str filename: Name for file in which to write function names.
-        :param Sequence[callable] functions: Functions on which this pipeline
-            is to operate (i.e., the functions for which name should be
-            written to output file).
-        """
-        # Set instance-specific variables.
-        self.name_output_file = filename
-        self.functions = functions
-        # Get the stages() benefit of superclass extension.
-        super(FunctionNameWriterPipeline, self).__init__(
-                name=name, outfolder=outfolder)
-
-    def write_name(self, func):
-        """
-        Write the name of a function to this pipeline's output file.
-
-        :param callable func: Name of function to write to the output file.
-        """
-        outpath = os.path.join(self.outfolder, self.name_output_file)
-        with open(outpath, 'a') as f:
-            f.write(func.__name__ + os.linesep)
-
-    def run(self):
-        """ Start with clean output file, then use superclass method. """
-        # Ensure that we start with a clean file since the nature of the
-        # operations performed (sequential file writes) creates desire to
-        # open output file in append mode rather than write mode.
-        output_file = os.path.join(self.outfolder, self.name_output_file)
-        if os.path.exists(output_file):
-            os.unlink(output_file)
-        super(FunctionNameWriterPipeline, self).run()
-
-
-    def stages(self):
-        """ Sequence of operations to perform. """
-        return [Stage(self.write_name, (f, ), name=f.__name__)
-                for f in self.functions]
-
-
-# Functions and fixtures
-
-def _get_read_aligner(outfolder):
-    """ Create a dummy 'read aligner' pipeline. """
-    return FunctionNameWriterPipeline(
-            "read-aligner", outfolder,
-            READ_ALIGNER_FILENAME, [merge_input, qc, align_reads])
-
-
-def _get_peak_caller(outfolder):
-    """ Create a dummy 'peak caller' pipeline. """
-    return FunctionNameWriterPipeline(
-        "peak-caller", outfolder,
-        PEAK_CALLER_FILENAME, [align_reads, call_peaks])
-
-def _get_pipeline(name, outfolder):
-    if name == read_aligner.__name__:
-        return _get_read_aligner(outfolder)
-    elif name == call_peaks.__name__:
-        return _get_peak_caller(outfolder)
-    else:
-        raise ValueError("Unknown pipeline request: '{}'".format(name))
-
-
-@pytest.fixture
-def read_aligner(tmpdir):
-    """ Provide test case with a read aligner pipeline instance. """
-    return _get_read_aligner(outfolder=tmpdir.strpath)
-
-
-@pytest.fixture
-def peak_caller(tmpdir):
-    """ Provide test case with a 'PeakCaller' pipeline instance. """
-    return _get_peak_caller(outfolder=tmpdir.strpath)
 
 
 
@@ -132,7 +18,7 @@ def test_pipeline_checkpoint_respect_sensitivity_checkpoint_perspective(
     """ Pipeline can skip past its stage(s) for which checkpoint exists. """
 
     # Create the pipeline.
-    pipeline = _get_pipeline(pl_name, tmpdir.strpath)
+    pipeline = get_pipeline(pl_name, tmpdir.strpath)
 
     # Negative control to start test, that we have no checkpoint files.
     assert [] == fetch_checkpoint_files(pipeline.manager)
@@ -196,7 +82,7 @@ def test_pipeline_checkpoint_sensitivity_effect_perspective(pl_name, tmpdir):
     """ The pipeline skips execution of stages with extant checkpoint. """
 
     # Create the pipeline, then check creation of output file.
-    pipeline = _get_pipeline(pl_name, tmpdir.strpath)
+    pipeline = get_pipeline(pl_name, tmpdir.strpath)
     output_file = os.path.join(pipeline.outfolder, pipeline.name_output_file)
     assert not os.path.exists(output_file)
     pipeline.run()
